+++ frontend/src/styles/reservationStyled.ts
